GTK: set chrome button xthickness to 2, regardless of theme.
In the upper chrome, we strongarm the vertical sizing of the buttons' containers, so we needn't worry about ythickness. The xthickness however affects the horizontal spacing of the buttons. This is true of both chrome theme mode and gtk+ theme mode.

Override the xthickness of the theme even in GTK+ mode because our UI is non-standard enough that it looks weird even there when the xthickness is too high.

BUG=46393
TEST=visual

Review URL: http://codereview.chromium.org/2800006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@50024 0039d316-1c4b-4281-b951-d872f2087c98